% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swimmerplot.R
\name{generate_tooltip}
\alias{generate_tooltip}
\title{Generate tooltips for interactive plot elements}
\usage{
generate_tooltip(dataset, tooltip_vector)
}
\arguments{
\item{dataset}{`[data.frame]` A data frame containing the variables to be used in tooltips}

\item{tooltip_vector}{`[character(n)]` A named character vector where names are labels and values are column names}
}
\value{
A character vector of HTML-formatted tooltip strings
}
\description{
Generate tooltips for interactive plot elements
}
\keyword{internal}
