% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_swimmerplot.R
\name{mock_swimmerplot}
\alias{mock_swimmerplot}
\title{Mock app for the swimmerplot module}
\usage{
mock_swimmerplot()
}
\description{
This function creates a standalone Shiny application that demonstrates the
swimmerplot module with example data from the pharmaversesdtm package.
}
\examples{
\dontrun{
mock_swimmerplot()
}

}
