% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_swimmerplot_mm.R
\name{mock_swimmerplot_mm}
\alias{mock_swimmerplot_mm}
\title{Run an example swimmer plot module in the module manager}
\usage{
mock_swimmerplot_mm()
}
\description{
Launches an example app that shows a swimmer plot module integrated in the
module manager surface. Displays data prepared from the pharmaversesdtm package.
}
\examples{
\dontrun{
mock_swimmerplot_mm()
}

}
