% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_swimmerplot.R
\name{swimmerplot_UI}
\alias{swimmerplot_UI}
\title{Swimmer Plot Module UI}
\usage{
swimmerplot_UI(
  id,
  group_by_vars = NULL,
  sort_by_vars = NULL,
  jumping_enabled = FALSE
)
}
\arguments{
\item{id}{`[character(1)]` Unique shiny ID. Must match the ID provided to [swimmerplot_server()].}

\item{group_by_vars}{`[character(n)]` Variables available for grouping subjects.}

\item{sort_by_vars}{`[character(n)]` Variables available for sorting subjects.}

\item{jumping_enabled}{`[logical(1)]` Whether clicking on a subject should enable navigation to detail view.}
}
\value{
Shiny UI.
}
\description{
(For use outside of the DaVinci framework)\cr
Places the Swimmer Plot module UI at the call site of this function. A matching call to [swimmerplot_server()]
is necessary.\cr
}
\seealso{
[mod_swimmerplot()] and [swimmerplot_server()]
}
